feat: result画面が寂しかったためmapを追加した
